B2, B3, B8 Figma Tasks (#1150)
* B2, B3, B8 Figma Tasks

fixes #1080
fixes #1081
fixes #1084
Fixes B2, B3, B8 of Figma tasks:
- added chips with functionality
- added pagination
- added totals for open vulnerabilities
- fixed styling for Most common ports
- fixed link in the See All button of Open Vuln card
- Adjusted the color for vulnerabilities with a null severity (e.g. hibp). This is not in figma, should I change this?

* small updates

- fixed typos
- added severity to dnstwist, hibp vulns

* lint

* Added # of MAX_RESULTS for stats

Reason why not all of the vulnerabilities were being displayed. I set it to 1000 as a draft. Let me know if you think it should be greater/less

* update num of vulns shown

Only show the top 50 domains with the most vulnerabilities (for the Open vulnerabilities by Domain chart)

* add low severity to hibp

* Added note to Open Vuln chart

Added note that we are only showing top 50 domains. Also, adjusted sizes and margins to do so